Move `AlgorithmConfig` to outer package
