+++ common/src/main/java/org/astraea/common/balancer/BalancerConsole.java
@@ -23,7 +23,6 @@ import java.util.concurrent.CompletionStage;
 import java.util.function.Function;
 import org.astraea.common.EnumInfo;
 import org.astraea.common.admin.Admin;
-import org.astraea.common.balancer.algorithms.AlgorithmConfig;
 import org.astraea.common.balancer.executor.RebalancePlanExecutor;
 
 /** Offer a uniform interface to schedule/manage/execute balance plan to an actual Kafka cluster. */
