+++ common/src/main/java/org/astraea/common/balancer/BalancerConsoleImpl.java
@@ -40,7 +40,6 @@ import org.astraea.common.admin.ClusterBean;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.NodeInfo;
 import org.astraea.common.admin.Replica;
-import org.astraea.common.balancer.algorithms.AlgorithmConfig;
 import org.astraea.common.balancer.executor.RebalancePlanExecutor;
 import org.astraea.common.cost.NoSufficientMetricsException;
 import org.astraea.common.metrics.collector.MetricCollector;
