+++ common/src/main/java/org/astraea/common/balancer/algorithms/GreedyBalancer.java
@@ -27,6 +27,7 @@ import java.util.function.Supplier;
 import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterInfo;
+import org.astraea.common.balancer.AlgorithmConfig;
 import org.astraea.common.balancer.Balancer;
 import org.astraea.common.balancer.tweakers.ShuffleTweaker;
 import org.astraea.common.cost.ClusterCost;
