+++ common/src/main/java/org/astraea/common/balancer/algorithms/SingleStepBalancer.java
@@ -22,6 +22,7 @@ import java.util.TreeSet;
 import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterInfo;
+import org.astraea.common.balancer.AlgorithmConfig;
 import org.astraea.common.balancer.Balancer;
 import org.astraea.common.balancer.tweakers.ShuffleTweaker;
