+++ common/src/test/java/org/astraea/common/balancer/algorithms/AlgorithmConfigTest.java
@@ -18,6 +18,7 @@ package org.astraea.common.balancer.algorithms;
 
 import org.astraea.common.admin.ClusterBean;
 import org.astraea.common.admin.ClusterInfo;
+import org.astraea.common.balancer.AlgorithmConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
