+++ common/src/test/java/org/astraea/common/balancer/algorithms/GreedyBalancerTest.java
@@ -23,6 +23,7 @@ import java.util.stream.IntStream;
 import org.astraea.common.Configuration;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterBean;
+import org.astraea.common.balancer.AlgorithmConfig;
 import org.astraea.common.balancer.FakeClusterInfo;
 import org.astraea.common.cost.DecreasingCost;
 import org.astraea.common.metrics.BeanQuery;
